Changed the corn/rice/oats section
Returned it to the text boxes I had created originally. I tried many
combinations of putting the class tag on the figure, the image and the
fig caption and tried a lot of different combinations of the
float/clear tags to get the bottom two pictures to line up correctly
but couldn’t figure it out. The only way I could get it to work was
like this so hopefully that’s ok!